# Conflicts:
#	androidHyperskillApp/src/main/java/org/hyperskill/app/android/core/injection/AndroidAppComponentImpl.kt
#	shared/src/commonMain/kotlin/org/hyperskill/app/core/injection/AppGraph.kt
#	shared/src/commonMain/kotlin/org/hyperskill/app/profile/domain/model/Profile.kt
#	shared/src/iosMain/kotlin/org/hyperskill/app/core/injection/AppGraphImpl.kt

+++ shared/src/commonMain/kotlin/org/hyperskill/app/profile/domain/model/Profile.kt
@@ -98,5 +98,5 @@ data class Profile(
     @SerialName("isic_name")
     val isicName: String,
     @SerialName("isic_status")
-    val isicStatus: Int
+    val isicStatus: Long
 )
